requirements.txt
torch>=2.0.0
transformers>=4.40.0
datasets>=2.19.0
accelerate>=0.30.0
scikit-learn>=1.4.0
gradio>=4.0.0
pyyaml>=6.0
rich>=13.0.0
numpy>=1.26.0
